// types/character.ts
export default interface Character {
  url: URL
  name: string
  gender: string
  culture: string
  born: string
  died: string
  titles: string[]
  aliases: string[]
  father: string
  mother: string
  spouse: string
  allegiances: string[]
  books: string[]
  povBooks: number[]
  tvSeries: string[]
  playedBy: string[]
}